+++ packages/moonstone/Picker/Picker.js
@@ -8,6 +8,8 @@
 import kind from '@enact/core/kind';
 import React from 'react';
 
+import {MarqueeController} from '../Marquee';
+
 import PickerCore from './PickerCore';
 import PickerItem from './PickerItem';
 import SpottablePicker from './SpottablePicker';
@@ -169,7 +171,12 @@ const PickerBase = kind({
  * @ui
  * @public
  */
-const Picker = SpottablePicker(PickerBase);
+const Picker = MarqueeController(
+	{startOnFocus: true},
+	SpottablePicker(
+		PickerBase
+	)
+);
 
 export default Picker;
 export {Picker, PickerBase};