+++ packages/moonstone/Scroller/Scrollbar.js
@@ -1,6 +1,7 @@
 import classNames from 'classnames';
 import React, {Component, PropTypes} from 'react';
 import ri from '@enact/ui/resolution';
+import Holdable from '@enact/ui/Holdable';
 import Spotlight from '@enact/spotlight';
 import {startJob, stopJob} from '@enact/core/jobs';
 import {contextTypes} from '@enact/i18n/I18nDecorator';
@@ -8,6 +9,8 @@ import {contextTypes} from '@enact/i18n/I18nDecorator';
 import IconButton from '../IconButton';
 
 import css from './Scrollbar.less';
+
+const HoldableIconButton = Holdable({endHold: 'onLeave'}, IconButton);
 
 const
 	verticalProperties = {
@@ -31,6 +34,7 @@ const
 		)
 	},
 	autoHideDelay = 200,
+	nop = () => {},
 	minThumbSize = ri.scale(4),
 	selectIcon = (isPrev) => (isVertical, rtl) => {
 		if (isVertical) {
@@ -90,9 +94,9 @@ class Scrollbar extends Component {
 	static contextTypes = contextTypes
 
 	static defaultProps = {
-		vertical: true,
-		onNextScroll: () => {},
-		onPrevScroll: () => {}
+		onNextScroll: nop,
+		onPrevScroll: nop,
+		vertical: true
 	}
 
 	autoHide = true
@@ -245,12 +249,12 @@ class Scrollbar extends Component {
 
 		return (
 			<div ref={this.initContainerRef} className={scrollbarClassNames}>
-				<IconButton backgroundOpacity="transparent" small disabled={prevButtonDisabled} className={prevButtonClass} onClick={clickPrevHandler}>
+				<HoldableIconButton backgroundOpacity="transparent" small disabled={prevButtonDisabled} className={prevButtonClass} onClick={clickPrevHandler} onHoldPulse={clickPrevHandler}>
 					{prevIcon}
-				</IconButton>
-				<IconButton backgroundOpacity="transparent" small disabled={nextButtonDisabled} className={nextButtonClass} onClick={clickNextHandler}>
+				</HoldableIconButton>
+				<HoldableIconButton backgroundOpacity="transparent" small disabled={nextButtonDisabled} className={nextButtonClass} onClick={clickNextHandler} onHoldPulse={clickNextHandler}>
 					{nextIcon}
-				</IconButton>
+				</HoldableIconButton>
 				<div ref={this.initThumbRef} className={thumbClass} />
 			</div>
 		);