+++ packages/sampler/stories/qa-stories/Popup.js
@@ -1,0 +1,43 @@
+import Button, {ButtonBase} from '@enact/moonstone/Button';
+import Popup, {PopupBase} from '@enact/moonstone/Popup';
+import React from 'react';
+import {storiesOf, action} from '@kadira/storybook';
+import {withKnobs, boolean, text, select} from '@kadira/storybook-addon-knobs';
+
+Button.propTypes = Object.assign({}, ButtonBase.propTypes, Button.propTypes);
+Button.defaultProps = Object.assign({}, ButtonBase.defaultProps, Button.defaultProps);
+Button.displayName = 'Button';
+
+Popup.propTypes = Object.assign({}, PopupBase.propTypes, Popup.propTypes);
+Popup.defaultProps = Object.assign({}, PopupBase.defaultProps, Popup.defaultProps);
+
+storiesOf('Popup')
+	.addDecorator(withKnobs)
+	.addWithInfo(
+		'using spotlightRestrict',
+		() => (
+			<div>
+				<p>
+					The contents of the popup below should contain the only controls that can be
+					navigated to using 5-way. This is because the popup is using a `spotlightRestrict`
+					value of `self-only`. If the value changes to `self-first`, the other panel controls
+					can receive focus, but priority will be given to controls within the popup first. If
+					the value changes to `none`, there is no priority.
+				</p>
+				<Button>Button</Button>
+				<Popup
+					open={boolean('open', true)}
+					noAnimation={boolean('noAnimation', false)}
+					noAutoDismiss={boolean('noAutoDismiss', false)}
+					onClose={action('onClose')}
+					showCloseButton={boolean('showCloseButton', true)}
+					spotlightRestrict={select('spotlightRestrict', ['none', 'self-first', 'self-only'], 'self-only')}
+				>
+					<div>{text('children', 'Hello Popup')}</div>
+					<br />
+					<Button>Button</Button>
+				</Popup>
+			</div>
+		)
+	);
+