+++ packages/moonstone/ContextualPopupDecorator/ContextualPopupDecorator.js
@@ -6,16 +6,20 @@
  * @module moonstone/ContextualPopupDecorator
  */
 
+import {forward} from '@enact/core/handle';
 import {hoc} from '@enact/core';
 import ri from '@enact/ui/resolution';
 import FloatingLayer from '@enact/ui/FloatingLayer';
 import {contextTypes} from '@enact/i18n/I18nDecorator';
+import Spotlight, {SpotlightContainerDecorator, spotlightDirections} from '@enact/spotlight';
 import React, {PropTypes} from 'react';
 
 import {ContextualPopup} from './ContextualPopup';
 import css from './ContextualPopupDecorator.less';
 
 const defaultConfig = {};
+const ContextualPopupContainer = SpotlightContainerDecorator({preserveId: true}, ContextualPopup);
+const depress = 'onKeyDown';
 
 /**
  * {@link moonstone/ContextualPopupDecorator.ContextualPopupDecorator} is a Higher-order Component
@@ -28,6 +32,7 @@ const defaultConfig = {};
  * @public
  */
 const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
+	const forwardDepress = forward(depress);
 
 	return class extends React.Component {
 		static displayName = 'ContextualPopupDecorator'
@@ -36,7 +41,8 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			super(props);
 			this.state = {
 				arrowPosition: {top: 0, left: 0},
-				containerPosition: {top: 0, left: 0}
+				containerPosition: {top: 0, left: 0},
+				containerId: Spotlight.add()
 			};
 
 			this.overflow = {};
@@ -66,7 +72,9 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			direction: PropTypes.oneOf(['up', 'down', 'left', 'right']),
 
 			/**
-			 * A function to be run when ContextualPopup closes.
+			 * A function to be run when either the close button is clicked or spotlight focus
+			 * moves outside the boundary of the popup. Setting `spotlightRestrict` to `'self-only'`
+			 * will prevent Spotlight focus from leaving the popup.
 			 *
 			 * @type {Function}
 			 * @public
@@ -83,7 +91,7 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			open: PropTypes.bool,
 
 			/**
-			 * Classname pass to the popup. You may set width and heights of the popup with it.
+			 * Classname to pass to the popup. You may set width and height of the popup with it.
 			 *
 			 * @type {String}
 			 * @public
@@ -97,7 +105,17 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * @public
 			 * @default false
 			 */
-			showCloseButton : PropTypes.bool
+			showCloseButton : PropTypes.bool,
+
+			/**
+			 * Restricts or prioritizes navigation when focus attempts to leave the popup. It
+			 * can be either `'none'`, `'self-first'`, or `'self-only'`.
+			 *
+			 * @type {String}
+			 * @default 'self-first'
+			 * @public
+			 */
+			spotlightRestrict: PropTypes.oneOf(['none', 'self-first', 'self-only'])
 		}
 
 		static contextTypes = contextTypes
@@ -105,7 +123,8 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		static defaultProps = {
 			direction: 'down',
 			open: false,
-			showCloseButton : false
+			showCloseButton: false,
+			spotlightRestrict: 'self-first'
 		}
 
 		componentWillReceiveProps (nextProps) {
@@ -113,6 +132,18 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				this.adjustedDirection = nextProps.direction;
 				this.setContainerPosition();
 			}
+		}
+
+		componentDidUpdate (prevProps) {
+			if (this.props.open && !prevProps.open) {
+				this.spotPopupContent();
+			} else if (!this.props.open && prevProps.open) {
+				Spotlight.focus();
+			}
+		}
+
+		componentWillUnmount () {
+			Spotlight.remove(this.state.containerId);
 		}
 
 		getContainerPosition (containerNode, clientNode) {
@@ -265,13 +296,40 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			this.clientNode = node;
 		}
 
+		handleKeyDown = (ev) => {
+			const {onClose} = this.props;
+			const direction = spotlightDirections[ev.keyCode];
+
+			if (direction) {
+				// prevent default page scrolling
+				ev.preventDefault();
+				// stop propagation to prevent default spotlight behavior
+				ev.stopPropagation();
+
+				// if focus has changed
+				if (Spotlight.move(direction)) {
+
+					// if current focus is not within the popup's container, issue the `onClose` event
+					if (!this.containerNode.contains(document.activeElement) && onClose) {
+						onClose(ev);
+					}
+				}
+			}
+
+			forwardDepress(ev, this.props);
+		}
+
+		spotPopupContent = () => {
+			Spotlight.focus(this.state.containerId);
+		}
+
 		render () {
-			const {showCloseButton, popupComponent: PopupComponent, popupClassName, open, onClose, ...rest} = this.props;
+			const {showCloseButton, popupComponent: PopupComponent, popupClassName, open, onClose, spotlightRestrict, ...rest} = this.props;
 
 			return (
 				<div className={css.contextualPopupDecorator}>
 					<FloatingLayer open={open} scrimType="none" onDismiss={onClose}>
-						<ContextualPopup
+						<ContextualPopupContainer
 							className={popupClassName}
 							showCloseButton={showCloseButton}
 							onCloseButtonClick={onClose}
@@ -279,9 +337,12 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 							arrowPosition={this.state.arrowPosition}
 							containerPosition={this.state.containerPosition}
 							containerRef={this.getContainerNode}
+							containerId={this.state.containerId}
+							spotlightRestrict={spotlightRestrict}
+							onKeyDown={this.handleKeyDown}
 						>
 							<PopupComponent />
-						</ContextualPopup>
+						</ContextualPopupContainer>
 					</FloatingLayer>
 					<div ref={this.getClientNode}>
 						<Wrapped {...rest} />