+++ packages/moonstone/Marquee/MarqueeController.js
@@ -1,3 +1,4 @@
+import {forward} from '@enact/core/handle';
 import hoc from '@enact/core/hoc';
 import React from 'react';
 
@@ -35,7 +36,8 @@ const contextTypes = {
 
 	/**
 	 * Called by Marquee instances when marqueeing is started (e.g. when focusing a Marquee
-	 * set to `marqueeOn='focus'`)
+	 * set to `marqueeOn='focus'`). If the Marquee instance should not or does not need to marquee,
+	 * the function can return `true` to mark itself complete.
 	 *
 	 * @type {Function}
 	 * @memberof moonstone/Marquee.Marquee.contextTypes
@@ -51,6 +53,27 @@ const contextTypes = {
 	unregister: React.PropTypes.func
 };
 
+
+/**
+ * Default configuration parameters for {@link moonstone/Marquee.MarqueeController}
+ *
+ * @type {Object}
+ * @memberof moonstone/Marquee.MarqueeController
+ * @hocconfig
+ */
+const defaultConfig = {
+	/**
+	 * When `true`, any `onFocus` events that bubble to the controller will start the contained
+	 * Marquee instances. This is useful when a component contains Marquee instances that need to be
+	 * started with sibling components are focused.
+	 *
+	 * @type {Boolean}
+	 * @default false
+	 * @memberof moonstone/Marquee.MarqueeController.defaultConfig
+	 */
+	startOnFocus: false
+};
+
 /**
  * {@link moonstone/Marquee.MarqueeController} is a Higher-order Component which will synchronize
  * contained Marquee's.
@@ -60,7 +83,10 @@ const contextTypes = {
  * @hoc
  * @public
  */
-const MarqueeController = hoc((config, Wrapped) => {
+const MarqueeController = hoc(defaultConfig, (config, Wrapped) => {
+	const {startOnFocus} = config;
+	const forwardBlur = forward('onBlur');
+	const forwardFocus = forward('onFocus');
 
 	return class extends React.Component {
 		static displayName = 'MarqueeController'
@@ -155,6 +181,22 @@ const MarqueeController = hoc((config, Wrapped) => {
 			}
 		}
 
+		/*
+		 * Handler for the focus event
+		 */
+		handleFocus = (ev) => {
+			this.dispatch('start');
+			forwardFocus(ev, this.props);
+		}
+
+		/*
+		 * Handler for the blur event
+		 */
+		handleBlur = (ev) => {
+			this.dispatch('stop');
+			forwardBlur(ev, this.props);
+		}
+
 		/**
 		 * Invokes the `action` handler for each synchronized component except the invoking
 		 * `component`.
@@ -168,7 +210,13 @@ const MarqueeController = hoc((config, Wrapped) => {
 			this.controlled.forEach((controlled) => {
 				const {component: controlledComponent, [action]: handler} = controlled;
 				if (component !== controlledComponent && typeof handler === 'function') {
-					handler.call(controlledComponent);
+					const complete = handler.call(controlledComponent);
+
+					// Returning `true` from a start request means that the marqueeing is
+					// unnecessary and is therefore complete
+					if (action === 'start' && complete) {
+						controlled.complete = true;
+					}
 				}
 			});
 		}
@@ -205,8 +253,18 @@ const MarqueeController = hoc((config, Wrapped) => {
 		}
 
 		render () {
+			let props = this.props;
+
+			if (startOnFocus) {
+				props = {
+					...this.props,
+					onBlur: this.handleBlur,
+					onFocus: this.handleFocus
+				};
+			}
+
 			return (
-				<Wrapped {...this.props} />
+				<Wrapped {...props} />
 			);
 		}
 	};