+++ packages/moonstone/VirtualList/VirtualList.js
@@ -10,7 +10,7 @@
 import kind from '@enact/core/kind';
 import React, {PropTypes} from 'react';
 
-import VirtualListBase from './VirtualListBase';
+import VirtualListBase, {gridListItemSizeShape} from './VirtualListBase';
 
 /**
  * {@link moonstone/VirtualList.VirtualList} is a VirtualList with Moonstone styling.
@@ -64,10 +64,10 @@ const VirtualGridList = kind({
 		 * <VirtualGridList itemSize={{minWidth: ri.scale(180), minHeight: ri.scale(270)}}/>
 		 * ```
 		 *
-		 * @type {Object}
+		 * @type {moonstone/VirtualList.gridListItemSizeShape}
 		 * @public
 		 */
-		itemSize: PropTypes.object.isRequired
+		itemSize: gridListItemSizeShape.isRequired
 	},
 
 	render: (props) => <VirtualListBase {...props} pageScroll />