+++ packages/ui/FloatingLayer/FloatingLayer.js
@@ -121,19 +121,13 @@ class FloatingLayerBase extends React.Component {
 		this.closeFloatingLayer();
 	}
 
-	handleClick = (ev) => {
+	handleClick = () => {
 		if (!this.props.noAutoDismiss && this.props.open && this.props.onDismiss) {
-			const scrim = this.floatLayer.querySelector('[data-ui-scrim]');
-
-			if (scrim) {
-				if (scrim.contains(ev.target)) {
-					this.props.onDismiss();
-				}
-			} else if (!this.floatLayer.contains(ev.target)) {
-				this.props.onDismiss();
-			}
+			this.props.onDismiss();
 		}
 	}
+
+	stopPropagation = (ev) => ev.nativeEvent.stopImmediatePropagation()
 
 	closeFloatingLayer () {
 		if (this.node) {
@@ -176,8 +170,8 @@ class FloatingLayerBase extends React.Component {
 		this.floatLayer = ReactDOM.unstable_renderSubtreeIntoContainer(
 			this,
 			<div {...rest}>
-				{scrimType !== 'none' ? <Scrim type={scrimType} data-ui-scrim /> : null}
-				{children}
+				{scrimType !== 'none' ? <Scrim type={scrimType} onClick={this.handleClick} /> : null}
+				{React.cloneElement(children, {onClick: this.stopPropagation})}
 			</div>,
 			node
 		);
@@ -187,7 +181,9 @@ class FloatingLayerBase extends React.Component {
 				onOpen();
 			}
 
-			on('click', this.handleClick);
+			if (scrimType === 'none') {
+				on('click', this.handleClick);
+			}
 		}
 	}