+++ packages/moonstone/CHANGELOG.md
@@ -6,6 +6,12 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Added
 
+- `autoClose` and `lockBottom` properties to `ExpandableItem`
+- `noAutoClose` and `noLockBottom` properties to `ExpandableList`
+
+### Changed
+
+- `autoClose` property of `moonstone/ExpandableList` to be `closeOnSelect` to disambiguate it from the added `autoClose` on 5-way up
 - `RadioDecorator` and `RadioControllerDecorator` components in `ui/RadioDecorator` module
 - Focus/blur handling to `moonstone/MarqueeController` to start and stop synchronized `moonstone/Marquee` components